Move all ratings test files into selenium-testing
