+++ selenium-test/setup.sh
